+++ packages/transaction-pay-controller/src/helpers/QuoteRefresher.ts
@@ -2,14 +2,13 @@ import type { TransactionMeta } from '@metamask/transaction-controller';
 import { createModuleLogger } from '@metamask/utils';
 import { noop } from 'lodash';
 
-import { TransactionPayStrategy } from '../constants';
-import { projectLogger } from '../logger';
 import type {
-  AccountSupports7702Callback,
   TransactionPayControllerMessenger,
   TransactionPayControllerState,
-  UpdateTransactionDataCallback,
-} from '../types';
+} from '..';
+import { TransactionPayStrategy } from '../constants';
+import { projectLogger } from '../logger';
+import type { UpdateTransactionDataCallback } from '../types';
 import { refreshQuotes } from '../utils/quotes';
 
 const CHECK_INTERVAL = 1000; // 1 Second
@@ -31,24 +30,19 @@ export class QuoteRefresher {
 
   readonly #updateTransactionData: UpdateTransactionDataCallback;
 
-  readonly #accountSupports7702: AccountSupports7702Callback;
-
   constructor({
     getStrategies,
     messenger,
     updateTransactionData,
-    accountSupports7702,
   }: {
     getStrategies: (transaction: TransactionMeta) => TransactionPayStrategy[];
     messenger: TransactionPayControllerMessenger;
     updateTransactionData: UpdateTransactionDataCallback;
-    accountSupports7702: AccountSupports7702Callback;
   }) {
     this.#getStrategies = getStrategies;
     this.#messenger = messenger;
     this.#isRunning = false;
     this.#isUpdating = false;
-    this.#accountSupports7702 = accountSupports7702;
     this.#updateTransactionData = updateTransactionData;
 
     messenger.subscribe(
@@ -87,7 +81,6 @@ export class QuoteRefresher {
         this.#messenger,
         this.#updateTransactionData,
         this.#getStrategies,
-        this.#accountSupports7702,
       );
     } catch (error) {
       log('Error refreshing quotes', error);